% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Show_Bots.R
\name{Show_Bots}
\alias{Show_Bots}
\title{Show_Bots}
\usage{
Show_Bots(Info_Loc = NULL)
}
\arguments{
\item{Info_Loc}{Where the RDS file with your needed bot info is saved. Defaults to path.expand("~")}
}
\value{
Bot names that are available in the specified folder.
}
\description{
Show a list of your bots that are available. By default it will search for bots in path.expand("~").
Alternatively,
}
\examples{
\dontrun{
Show_bot()
}
}
